+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- New `number` condition.
+
 ## 2.7.0 - 2019-06-20
 
 ### Added